# servers/customer-services/docs/api-documentation.yaml
openapi: 3.1.0
info:
  title: InsightSerenity Customer Services API
  version: 1.0.0
  description: |
    # Customer Services API Documentation

    ## Overview
    The Customer Services API provides multi-tenant business services including:
    - User Management
    - Client Management
    - Project Management
    - Recruitment Services
    - Organization Management

    ## Authentication
    All endpoints require JWT Bearer token authentication.

    ## Multi-tenancy
    All requests must include the `X-Tenant-ID` header.
servers:
  - url: http://localhost:3001
    description: Development Server
  - url: https://api.insightserenity.com
    description: Production Server
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        200:
          description: Token refreshed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...